daemon/requirements.txt
cryptography==3.4.7
influxdb-client==1.18.0
pycryptodomex==3.10.1
requests==2.25.1
